test linguistic summarization methods

+++ src/main/java/data/Footballer.java
@@ -7,21 +7,21 @@ import lombok.Setter;
 @Setter
 public class Footballer {
 
-    private Integer id;
+    private double id;
     private String name;
-    private Double age; // wiek (very old, old, average, young, very young)
-    private Double potential; // (legendary, world class, professional, novice, amateur, beginner)
-    private Double value; // wartość rynkowa (very cheap, cheap, average, expensive, very expensive)
-    private Double height; // wzrost w cm (very high, high, medium, low, very low)
-    private Double weight; // waga w kg (heavy, semi-heavy, medium, welterweight, light)
-    private Double speed; // szybkość (legendary, world class, professional, novice, amateur, beginner)
-    private Double stamina; // wytrzymałość (legendary, world class, professional, novice, amateur, beginner)
-    private Double strength; // siła (legendary, world class, professional, novice, amateur, beginner)
-    private Double balance; // równowaga (legendary, world class, professional, novice, amateur, beginner)
-    private Double agility; // zwinność (legendary, world class, professional, novice, amateur, beginner)
+    private double age; // wiek (very old, old, average, young, very young)
+    private double potential; // (legendary, world class, professional, novice, amateur, beginner)
+    private double value; // wartość rynkowa (very cheap, cheap, average, expensive, very expensive)
+    private double height; // wzrost w cm (very high, high, medium, low, very low)
+    private double weight; // waga w kg (heavy, semi-heavy, medium, welterweight, light)
+    private double speed; // szybkość (legendary, world class, professional, novice, amateur, beginner)
+    private double stamina; // wytrzymałość (legendary, world class, professional, novice, amateur, beginner)
+    private double strength; // siła (legendary, world class, professional, novice, amateur, beginner)
+    private double balance; // równowaga (legendary, world class, professional, novice, amateur, beginner)
+    private double agility; // zwinność (legendary, world class, professional, novice, amateur, beginner)
 
-    public Footballer(Integer id, String name, Double age, Double potential, Double value, Double height,
-                      Double weight, Double speed, Double stamina, Double strength, Double balance, Double agility) {
+    public Footballer(int id, String name, int age, int potential, int value, double height,
+                      double weight, int speed, int stamina, int strength, int balance, int agility) {
         this.id = id;
         this.name = name;
         this.age = age;
